requirements.txt
langchain-openai
langchain-community
faiss-cpu
lxml~=6.0.2
beautifulsoup4~=4.14.3
tiktoken~=0.12.0
langchain~=1.2.7
python-dotenv~=1.2.1
playwright~=1.57.0
playwright-stealth